// src/enums/report.tutorial.enum.ts
export enum EReportTutorialTitle {
  STEP1 = '¿Encontraste alguna falla? No te preocupes, nosotros te apoyamos',
  STEP2 = 'Toma una foto para documentar la falla o incidencia',
  STEP3 = 'Haz una breve descripción del incidente e indicanos el lugar donde ocurrio',
  STEP4 = 'Listo el incidente será informado y pronto lo repararemos',
}

export enum EReportStepsTutorial {
  STEP1 = 1,
  STEP2 = 2,
  STEP3 = 3,
  STEP4 = 4,
}